% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_citescope.R
\name{run_app}
\alias{run_app}
\title{Launch the CITEscope Shiny app}
\usage{
run_app(workers = 4, upload_max_mb = 1024, future_max_gb = 15)
}
\arguments{
\item{workers}{Number of parallel workers (default: 4)}

\item{upload_max_mb}{Max single-file upload size in MB (default: 1024 = 1GB)}

\item{future_max_gb}{Max future globals size in GB (default: 15)}
}
\value{
shiny.appobj
}
\description{
Launch the CITEscope Shiny app
}
